# api/openapi.yaml
openapi: 3.0.3
info:
  title: FolioPath HTTP API
  version: 0.1.0
  description: |
    Authoritative HTTP contract through the FolioPath POST-MVP-5 scope,
    revision 2. Revision-1 image semantic operations are contract-frozen.
    Revision-2 controlled tag suggestions, video semantic search, and local
    anonymous face/person curation remain gated from production composition
    until INT-S1R2 is accepted.

    All business endpoints use an authenticated same-origin browser session.
    State-changing authenticated operations additionally require the
    session-bound `X-CSRF-Token`. First-run setup and login do not yet have a
    session, so they require a validated same-origin `Origin` header.

    IDs and cursors are opaque. Media content endpoints accept only resource
    IDs, never filesystem paths. Management responses may expose a library
    root relative to the allowed `/library` container mount and a server-built
    `/library/...` display label; no response may expose a host path.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
...
servers:
  - url: /
    description: Same-origin FolioPath server
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/ai/models:
    get:
      tags: [AI models]
      operationId: listAIModels
      summary: List reviewed installed models and their safe status
      description: |
        Returns only reviewed package metadata and opaque identifiers. It does
        not expose model filenames, container paths, host paths, source URLs,
        runtime errors, or unreviewed files found below `/models`.
      x-requirements: [FR-INT-015, FR-INT-017, FR-INT-020, NFR-INT-005, NFR-INT-008]
      x-error-codes:
        '401': [authentication_required, session_expired]
        '429': [rate_limited]
        '500': [internal_error]
      responses:
        '200':
          description: Installed reviewed model packages.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/ai/model-candidate-scans:
    post:
      tags: [AI models]
      operationId: scanAIModelCandidates
      summary: Scan the fixed read-only model directory
      description: |
        Scans only the server-configured `/models` boundary. The request has no
        path, URL, source, proxy, mirror, or filename field. Candidate IDs bind
        the resulting scan revision and expire when the source changes.
        Before scanning candidates, the server revalidates every installed
        managed/direct package against its reviewed manifest and updates only
        its availability state; active generation metadata and embeddings are
        preserved when a source is unavailable.
        Revision 1 deliberately has no online model-download operation.
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-017, FR-INT-018, NFR-INT-007, NFR-INT-008, NFR-INT-010]
      x-error-codes:
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '409': [model_source_unavailable]
        '429': [rate_limited]
        '500': [internal_error]
      responses:
        '200':
          description: A bounded safe snapshot of reviewed and rejected candidates.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelCandidateScan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/ai/model-candidates/{candidateId}/install:
    parameters:
      - $ref: '#/components/parameters/AIModelCandidateIDParameter'
    post:
      tags: [AI models]
      operationId: installAIModelCandidate
      summary: Install one candidate by managed copy or strict direct reference
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-017, FR-INT-018, FR-INT-019, NFR-INT-007, NFR-INT-009, NFR-INT-010]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '409': [model_candidate_stale, idempotency_conflict]
        '422': [model_incompatible, model_source_unsafe, insufficient_space]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallAIModelRequest'
      responses:
        '202':
          description: Verification and installation operation accepted.
          headers:
            Location:
              description: Relative URL of the model operation.
              schema:
                type: string
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/ai/models/{modelId}/activate:
    parameters:
      - $ref: '#/components/parameters/AIModelIDParameter'
    post:
      tags: [AI models]
      operationId: activateAIModel
      summary: Build and activate a new semantic generation
      description: |
        The old active model and generation remain authoritative until this operation succeeds.
        If-Match uses the selected model ID and its availabilityRevision in the form
        `"{modelId}-r{availabilityRevision}"`.
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-005, FR-INT-016, FR-INT-020, NFR-INT-004]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '409': [model_unavailable, idempotency_conflict]
        '412': [precondition_failed]
        '428': [precondition_required]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IfMatchHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '202':
          description: Activation operation accepted.
          headers:
            Location:
              description: Relative URL of the model operation.
              schema:
                type: string
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/ai/operations/{operationId}:
    parameters:
      - $ref: '#/components/parameters/AIOperationIDParameter'
    get:
      tags: [AI models]
      operationId: getAIOperation
      summary: Get a model or semantic operation
      x-requirements: [FR-INT-015, FR-INT-016, NFR-INT-002]
      x-error-codes:
        '401': [authentication_required, session_expired]
        '404': [ai_operation_not_found]
        '429': [rate_limited]
        '500': [internal_error]
      responses:
        '200':
          description: Current operation state.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelOperation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/ai/operations/{operationId}/cancel:
    parameters:
      - $ref: '#/components/parameters/AIOperationIDParameter'
    post:
      tags: [AI models]
      operationId: cancelAIOperation
      summary: Cooperatively cancel an active AI operation
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-016, NFR-INT-002]
      x-error-codes:
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '404': [ai_operation_not_found]
        '409': [ai_operation_already_finished]
        '412': [precondition_failed]
        '428': [precondition_required]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IfMatchHeader'
      responses:
        '202':
          description: Cancellation accepted or already being applied.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelOperation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: foliopath_session
      description: |
        High-entropy, opaque session cookie backed by a revocable server-side
        session. It is HttpOnly and SameSite=Strict; deployments using HTTPS
        set Secure. The database stores no replayable plaintext token.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Session-bound token returned by setup, login, and session responses.
        It is required together with `cookieAuth` on authenticated
        state-changing operations. It must never be placed in a URL.
  parameters:
...
    AIModelIDParameter:
      name: modelId
      in: path
      required: true
      description: Opaque reviewed model ID with no filename or path semantics.
      schema:
        $ref: '#/components/schemas/ResourceID'
    AIModelCandidateIDParameter:
      name: candidateId
      in: path
      required: true
      description: Opaque, expiring candidate ID bound to one fixed-directory scan revision.
      schema:
        $ref: '#/components/schemas/ResourceID'
    AIOperationIDParameter:
      name: operationId
      in: path
      required: true
      description: Opaque model or semantic operation ID.
      schema:
        $ref: '#/components/schemas/ResourceID'
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated opaque key retained for at least 24 hours. Use a new
        key for a different logical action. Reuse with a different request body
        fails with `idempotency_conflict`. The server stores only a fixed-size
        digest of this value and never writes the plaintext key to logs.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._~-]+$'
    IfMatchHeader:
      name: If-Match
      in: header
      required: true
      description: Strong ETag from the latest GET. Missing values fail with 428 and stale values with 412.
      schema:
        type: string
        minLength: 3
        maxLength: 256
...
  headers:
...
    RequestID:
      description: Server-generated opaque correlation ID. Client-supplied values are not trusted.
      schema:
        type: string
        pattern: '^req_[A-Za-z0-9_-]+$'
        maxLength: 128
...
    IdempotencyReplayed:
      description: True only when this representation is a replay of a previously completed idempotent request.
      schema:
        type: boolean
    ETag:
      description: |
        Strong opaque representation validator. For a Library it changes when
        any represented field changes, including scan-derived status/counts.
      schema:
        type: string
        example: '"fp-8b51cdd5"'
...
  responses:
    BadRequest:
      description: The request syntax, cursor, query combination, or encoded input is invalid.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: A valid administrator session is required or the session has expired.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: CSRF or same-origin validation failed. No protected action was executed.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested opaque resource ID does not identify a visible resource.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: |
        The request conflicts with durable state, such as closed setup,
        duplicate library name, overlapping root, active removal, offline
        source, terminal scan, or idempotency-key body mismatch.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PreconditionFailed:
      description: The supplied If-Match no longer identifies the current representation.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        ETag:
          $ref: '#/components/headers/ETag'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PreconditionRequired:
      description: A current If-Match validator is required.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: The JSON shape is valid but one or more values violate the accepted contract.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: A per-endpoint authentication, browse, scan, thumbnail, or media limit was reached.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected failure represented by a safe public code and request ID.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    NullableResourceID:
      type: string
      nullable: true
      description: Opaque stable resource identifier, or null when no related resource exists.
      minLength: 3
      maxLength: 128
      pattern: '^[A-Za-z][A-Za-z0-9_-]+$'
...
    Timestamp:
      type: string
      format: date-time
      description: UTC RFC 3339 timestamp.
...
    AIModelPage:
      type: object
      additionalProperties: false
      required: [items, activeModelId, revision]
      properties:
        items:
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/AIModel'
        activeModelId:
          $ref: '#/components/schemas/NullableResourceID'
        revision:
          type: integer
          format: int64
          minimum: 1
...
    AIModelCandidateScan:
      type: object
      additionalProperties: false
      required: [revision, scannedAt, candidates, truncated]
      properties:
        revision:
          type: integer
          format: int64
          minimum: 1
        scannedAt:
          $ref: '#/components/schemas/Timestamp'
        candidates:
          type: array
          maxItems: 64
          items:
            $ref: '#/components/schemas/AIModelCandidate'
        truncated:
          type: boolean
          description: True when the bounded scan found more entries than may be reported.
...
    InstallAIModelRequest:
      type: object
      additionalProperties: false
      required: [storageMode]
      properties:
        storageMode:
          type: string
          enum: [managed, direct]
    AIModelOperation:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - state
        - phase
        - completedItems
        - totalItems
        - revision
        - errorCode
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/ResourceID'
        kind:
          type: string
          enum:
            - model_install
            - model_activate
            - semantic_missing
            - semantic_rebuild
            - semantic_clear
            - tag_suggestion_missing
            - tag_suggestion_rebuild
            - tag_review_clear
            - video_semantic_missing
            - video_semantic_rebuild
            - face_missing
            - face_rebuild
            - face_derived_clear
            - face_manual_clear
        state:
          type: string
          enum: [queued, running, cancelling, succeeded, failed, cancelled]
        phase:
          type: string
          enum: [queued, scanning, verifying, copying, loading, building, detecting, clustering, validating, clearing, finalizing, completed]
        completedItems:
          type: integer
          format: int64
          minimum: 0
        totalItems:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        revision:
          type: integer
          format: int64
          minimum: 1
        errorCode:
          type: string
          nullable: true
          enum:
            - model_candidate_stale
            - model_incompatible
            - model_source_unavailable
            - model_source_unsafe
            - model_unavailable
            - vocabulary_unavailable
            - storyboard_not_ready
            - face_model_unavailable
            - insufficient_space
            - source_offline
            - source_missing
            - source_unreadable
            - operation_interrupted
            - cancelled
            - internal_error
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
...